fix(web): consolidate marketplace logo and correct public asset path
