+++ web/app/components/plugins/marketplace/home/__tests__/home-header.spec.tsx
@@ -102,16 +102,10 @@ describe('HomeHeader', () => {
     render(<HomeHeader isMarketplacePlatform />)
 
     const brandLink = screen.getByRole('link', { name: 'Dify Marketplace' })
-    const [lightLogo, darkLogo] = brandLink.querySelectorAll('img')
-    expect(lightLogo).toHaveAttribute('src', expect.stringContaining('dify-marketplace-logo.svg'))
-    expect(darkLogo).toHaveAttribute(
-      'src',
-      expect.stringContaining('dify-marketplace-logo-dark.svg'),
-    )
-    expect(lightLogo).toHaveAttribute('width', '141.761')
-    expect(lightLogo).toHaveAttribute('height', '16.386')
-    expect(darkLogo).toHaveAttribute('width', '141.761')
-    expect(darkLogo).toHaveAttribute('height', '16.386')
+    const logo = brandLink.querySelector('img')
+    expect(logo).toHaveAttribute('src', '/marketplace/dify-marketplace-logo.svg')
+    expect(logo).toHaveAttribute('width', '141.761')
+    expect(logo).toHaveAttribute('height', '16.386')
     expect(screen.queryByText('mainNav.marketplace')).not.toBeInTheDocument()
   })
