+++ web/app/components/plugins/marketplace/marketplace-logo.tsx
@@ -1,0 +1,14 @@
+import { basePath } from '@/utils/var'
+
+export function MarketplaceLogo() {
+  return (
+    <img
+      alt=""
+      aria-hidden
+      className="block h-[16.386px] w-[141.761px] max-w-none shrink-0 [html[data-theme=dark]_&]:brightness-0 [html[data-theme=dark]_&]:invert-91"
+      height="16.386"
+      src={`${basePath}/marketplace/dify-marketplace-logo.svg`}
+      width="141.761"
+    />
+  )
+}
